minor refactor and bug fixes
moved methods that update the list on the website, methods that update the files, methods that pull data from the website, and methods that update the client side stored list to different files to prevent future cyclical imports.

Also allows the creation of new files that include the animeId. The plan is to  get rid of accessing anime through their names and use exclusively their IDs.

Removed bug that crashed the program if an anime not on any anime lists tried to be created or updated.