fixed an error
when there would be a folder or file in the Data folder that's not a status, the "moveFiles" method would break because it reads the directories for the statuses. Instead the statuses are now hard coded. This commit also includes miscellaneous changes to other files such as adding imports or fixing some syntax errors.